Don't coerce the option default to a RankedValue. (#21599)
This is unnecessary (and I don't recall if or why it ever was necessary), and complicates switching to using value derivation history from the native parser. 

At the only site where we consumed this coerced value we immediately took its inner value, so we might as well just have that be the default kwarg all along.

Note that this coercion was only happening in the iterator over registration kwargs, which isn't actually used during option parsing anyway. So all this is only relevant to help generation.