Revert "Replace macOS 10.15 x86-64 self-hosted runner with 12 GH-hosted (#21417)" (#21623)
This reverts commit ae20980b59f529ba5bde97ded3aca712ddfbc75e / #21417 to
restore current-`main`/version 2.24.x to using our self-hosted macOS
10.15 runner, rather than the Github-hosted macOS 12 runner that is
being deprecated on 2024-12-03.

This ensures we'll be able to do the 2.24.0 release and patch releases
after that. This doesn't change the deprecation of macOS versions 10.15,
11 (in 2.23, #21326) and 12 (in 2.24, #21569). We'll be "unnecessarily"
building on 10.15 for 2.24 even though it's not actually supported. The
warnings/errors about the lack of support will still occur, though.

Once we branch for 2.24 and `main` switches over to 2.25, we can switch
this over to using the GitHub hosted macOS 13 runner (not 12).

See
https://github.com/pantsbuild/pants/issues/21333#issuecomment-2460869204
for more discussion.