Use option value derivation from native parser. (#21627)
The derivation (previously inaptly called "history" on the python side)
is displayed to users in option help. It shows what the option value
provided by each source type (config/env/flag) was, as well as the final
value (from the highest ranked source).

For scalar values this is relatively simple, as values can only be
overridden. For list and dict values this is more complicated,
as values from one source can be appended to by another source. 
So a final value (or any intermediate value) may come from multiple
sources.

The goal in this PR was to be compatible with the existing help
system tests, so we can be confident that we have equivalent
behavior. Once the legacy parser is gone we can consider
streamlining parts of this.