Use references to options Sources, instead of cloning. (#21629)
There are a small handful of `Source` instances that we keep 
cloning over and over again for every option value and 
derivation. In particular, the`Source` instances that represent
config files include the path to the config file, so this cloning
involves a lot of superfluous string allocation and copying.

Instead, we now use a reference with the lifetime of the 
OptionParser.

This is preparation work for passing string descriptors of
the source of each option value and derivation back to 
Python.