Test and release on newer macOS: 13 (x86-64) and 14 (arm64) (#21655)
This follows-up on the deprecation of support for older macOS versions
that happened in 2.24.x and earlier, by switching `main` (2.25.x) to be
building on macOS 13 for x86-64 (upgrading from 10.15) and macOS 14 for
arm64 (upgrading from 11). These particular versions are dictated by
Github's hosted runners: there's no arm64 macos-13 runners.

| version | branch | supported (x86-64) | built-on (x86-64)   | supported (arm64) | built-on (arm64) | refs                                        |        |
|---------|--------|--------------------|---------------------|-------------------|------------------|---------------------------------------------|--------|
| 2.22.x  | 2.22.x | 10.15              | 10.15 (self-hosted) | 11                | 11 (self)        | #21074                                      | #21326 |
| 2.23.x  | 2.23.x | 12                 | 10.15 (self)        | 12                | 11 (self)        | #21569 (also #21417 but reverted in #21623) |        |
| 2.24.x  | 2.24.x | 13                 | 10.15 (self)        | 14                | 11 (self)        | #21333, #21623                              |        |
| 2.25.x  | main   | 13                 | 13 (Github hosted)  | 14                | 14 (GH)          | #21333, #21413, this PR                     |        |

This is _preparation_ for scaling down our self-hosted macOS runners,
but we cannot do that yet. They're still used by:

- 2.24.x branches and earlier (in this repo)
- scie-pants' CI (https://github.com/pantsbuild/scie-pants/issues/421)

Thus, we at least need to wait a few months for the 2.24.x milestone to
be closed before we can reduce that spend.

(NB. despite the title, this doesn't _start_ testing on arm64 macOS: see
separate issue #20993.)

Fixes #21413
Fixes #21333