A global lock on side effects into the workspace. (Cherry-pick of #21698) (#21708)
Although Workspace.write_digest() is typically used in @goal_rules, 
and so is not invoked concurrently, there are a handful of cases where
we might need to write deeper inside the rule graph, and therefore
may run into concurrency issues.

One example is the bsp server, which may attempt to write build
byproducts into the workspace for the IDE to consume, and in some
cases may try and write the same files concurrently. Therefore we
force all workspace writes to run in a critical section.

In the common case of @goal_rule, the calls are sequential anyway,
and so there will be no lock contention. In the bsp case, locking is
required, and any contention is necessary.