generalize glob syntax for known third party docker images
In #21076 a subsystem was introduced to provide better warnings for
images that are "not found". It includes a method to either listing
out known images, or using `*` as a wildcard to assume "everything
else" that doesn't look like a target is an image.  This diff,
generalized the `*` syntax to allow expressing something like
`my.internal.registry.io/*` is the (only) source of non-target images.

ref #18346