do not forward vars files when running on terraform modules
When running on a module, `get_terraform_backend_and_vars` would detect all vars files.
(Because the vars files are typically specified as dependencies on the deployment, the module has none.
Inference then grabs all of them.)
This is almost certainly not desired,
especially since they will be passed when running against the deployment.