fix infinite loop due to panic from bare `.unwrap` of a `Result` in `create_digests` intrinsic (#21746)
While investigating https://github.com/pantsbuild/pants/issues/21709,
the investigation was obscured by the fact that the issue manifested as
a panic in the `create_digests` intrinsic. The error was in the `Err`
variant of a `Result` which the `create_digests` blindly called
`.unwrap()` on, which of course panics by design.

This panic resulted in the engine retrying the build graph node at issue
and hitting the panic again ... and again ... and again ...

The engine should really stop if it encounters a panic from a node
computation, but in the meantime, the `create_digests` should just
return errors in `Result` instances instead of panicking.