only emit a `mkdir` into remote execution wrapper for non-empty paths (#21753)
As reported in https://github.com/pantsbuild/pants/issues/21747, the
remote execution wrapper used to set up append-only cache directories
was trying to create an empty path as a directory. This occurred because
`Path::parent` returns an empty string (`Some("")`) if the path was not
absolute (i.e., the "root" was not `""` and not `"/"`).

Solution: Check for that condition and do not emit the `mkdir` in that
case.