Use OptionInfo type for option registration args/kwargs. (#21766)
The options registration system has for a long time passed around
the naked args and kwargs used at registration time.

However the more recent implementation of class field-based
declarative options introduced an `OptionsInfo` dataclass to
encapsulate that data. 

This PR spreads the use of `OptionsInfo` into the options
registration code. Advantages include:

- More succinct code.
- Proper type annotations.
- No confusion about when to use `*` or `**`.

Note that this change renames `OptionsInfo` to `OptionInfo`, 
since it represents the registration of a single option.
It also renames its fields to `args` and `kwargs`, since that
is the usage in most of the related code. Also, the name `flag_options`
was doubly confusing: A) A flag is just one aspect of an option, and 
B) "options" is overloaded. This refers the knobs you can set when 
registering an option, so using the word "options" for it is
brain-hurting.
I considered "knobs", but figured "kwargs" was just as good, and already
in use.