infer the PBS release tag from the user-supplied URLs (#21739)
Infer the PBS "release tag" for PBS releases supplied via the
`--python-build-standalone-known-python-versions` option from the given
URLs if those URLs conform to the naming convention used by the PBS
project.

The new advanced option
`--python-build-standalone-require-inferrable-release-tag` controls
whether Pants requires the tag to be inferrable. This option currently
defaults to `False`, but will be migrated to `True` in a future Pants
release.

Note: This PR does not implement explicitly providing the release tag
via `--python-build-standalone-known-python-versions` since changing how
the `|`-separated values are parsed would need to be done in a
backwards-compatible manner. That is something we may want to explore in
a future PR but this PR does not go there.

---------

Co-authored-by: Tom Dyas <tom@shoalshoftware.com>
Co-authored-by: Huon Wilson <wilson.huon@gmail.com>