make Platform class serializable for workunit-logger (#21792)
As described in #21773, having the Helm backend lint with the
workunit-logger results in json serialization errors. Giving `Platform`
a string representation resolves this (and anywhere else we may want to
"print" it.

Example workunit:
```
{"name": "pants.backend.helm.util_rules.tool.download_external_helm_plugin", "span_id": "bdaa19a5cd1ef797", "level": "DEBUG", "parent_id": "c8c137a5b78f02ea", "start_secs": 1734722271, "start_nanos": 296631921, "duration_secs": 0, "duration_nanos": 51846519, "description": "Download external Helm plugin", "metadata": {"name": "unittest", "version": "0.3.3", "platform": "linux_x86_64"}}
```

NOTE: I think this is a totally safe ~one line change, but admit it is
to a rather core class.

fixes #21773