Upgrade the default version of black from 23.12 to 24.8 (#21056)
This upgrades the default black version from 23.12.1 to 24.8.0. 

This results in "user-visible" **formatting changes**: we're ticking
over into a new year, and thus pick up the "2024" stable style in
24.1.0: https://black.readthedocs.io/en/stable/change_log.html#id18 .
Depending on the codebase, this potentially results in significant
reformatting (and does so for Pants).

To make it easier for people to upgrade, this retains a copy of the old
lockfile, that can be referenced via a `resources:` URI.

There's various other changes in the release notes, like support for new
syntax, bug fixes and performance improvements.

Because Pants uses its own black backend this PR also reformats the
world, according to the new style.

The commits are independently sensible:

1. Upgrade black itself
2. Adjust the flake8 rules due to a new conflict
(https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#e701-e704)
3. Reformat the whole world
4. Various fixes after review