Simplify OptionsBootstrapper (#21784)
Now that options parsing is done in Rust, much of the logic
in the OptionsBootstrapper was unneeded. This change
removes much of that defunct logic, particularly around
config file discovery, and creation of intermediate 
"bootstrap options".

Unfortunately this logic was still pretty entangled into the
bootstrap sequence, so this change is slightly intricate.