support `{chroot}` for program arguments in remote execution  (#21803)
As reported in https://github.com/pantsbuild/pants/issues/17519, remote
environments do not currently support the `{chroot}` replacement marker
supported by the local and Docker environments. With local and Docker
environments, Pants will replace any `{chroot}` marker with the absolute
path of the sandbox directory if it appears in a program argument or
environment variable.

This PR is a partial solution and implements `{chroot}` replacement for
prorgram arguments used in remote execution. It by extends the existing
wrapper script used for append-only caches to also support `{chroot}`
replacements.

This PR does not try to implement `{chroot}` replacement for environment
variables. That will need to come after I figure out a good way to do
that.