skip provider cache for `terraform init` if there is no lockfile
The worst concurrency bug with the Terraform provider cache
is when nonatomic moves result in incorrect hashes in the lockfile
which gets pulled into the Pants cache and then poisons it.
So we just don't use the provider cache in that case.

I think there's still the possibility that the nonatomic move triggers a problem with a module with a lockfile when the cache is being populated,
but at least that won't enter the cache.
I think that risk is lower since the files are fetched during `init` but used later (ex during `validate`),
by which time the copy operation has completed.