nfpm(bugfix): handle package targets w/ output_path
A package target with an output_path puts the packaged file
relative to the build_root instead of relative to the BUILD
file. So, we need to explicitly check for both the file_path
(relative to BUILD file) and the value (which might be
relative to the build_root) instead of just the file_path.