nfpm(bugfix): handle package targets w/ output_path (Cherry-pick of #21806) (#21808)
A package target with an output_path puts the packaged file relative to
the build_root instead of relative to the BUILD file. So, we need to
explicitly check for both the file_path (relative to BUILD file) and the
value (which might be relative to the build_root) instead of just the
file_path.

In debugging this, I ended up copying the plugin code into a separate
repo so I could figure out what was going wrong. It would have been
easier to debug if the sandbox contents were listed in the error message
that claims the src file was missing. So, this also expands the error
messages so future issues can be identified (hopefully) more simply.

Co-authored-by: Jacob Floyd <cognifloyd@gmail.com>