fix flaky failure in `export_integration_test.py` (Cherry-pick of #21756) (#21820)
`src/python/pants/backend/python/goals/export_integration_test.py` was
failing for me consistently in a PR of mine and on `main` locally. Not
sure why it passes elsewhere.

See this failure for example:
https://github.com/pantsbuild/pants/actions/runs/12306576175/job/34349676205?pr=21745#step:10:478

The test is checking whether a venv activation script has a certain
prompt with both a resolve name and the Python version. The test also
checks whether the quote character starting that string begins with a
double-quote character.

On my system for `main` and in the PR on CI, that quote character became
a single quote instead so the test fails.

Use a regex and allow either quote character.

Co-authored-by: Tom Dyas <tom@shoalsoftware.com>