Reimplement the ArgSplitter in Rust. (#21814)
The only functionality in the legacy ArgSplitter not
replicated in Rust is distinguishing between 
builtin/auxiliary goals and regular goals. This
can easily be done post-hoc, in Python, as 
that distinction is not required in order to correctly
split args (now that we don't assign flags to
scopes in the ArgSplitter).

The tests are ported directly from the relevant 
Python ones.

A future PR will actually wire this new ArgSplitter
up instead of the Python one.