Add plugin hook to nfpm backend so plugins can inject nfpm package field values (#21822)
With the nfpm backend, I need to inject some field values, like
`version` and `version_release` from an in-repo plugin. In particular, I
need to retrieve the `version` from a file in the repo, and the
`version_release` needs to be calculated using some API calls to the
remote package repository.

I could inject `version` using .pants.bootstrap as described here:
https://www.pantsbuild.org/blog/2024/04/27/simple-versioning-with-git-tags
However, looking up the value for `version_release` requires some of the
parametrized values of the package target's fields (like `platform`,
`arch`, and a custom field for distribution major version). So, I need a
way to dynamically inject some of these metadata values.