golang: allow downloads during instalation of gRPC protobuf plugins (#21614)
As described in https://github.com/pantsbuild/pants/issues/21529, the Go
rules are setting `GOPROXY=off` during the invocations of `go install`
to install the gRPC protobuf plugins because that is the default for
`GoSdkProcess`. The plugin builds are failing because `GOPROXY=off`
prevents `go` from contacting the Go module proxy to learn about module
deprecations. (Unlike other uses of `GoSdkProcess`, the gRPC protobuf
plugins are relying on `go install` to do a full build of the plugin
sources, which is unlike the use of `GoSdkProcess` in the rest of the Go
backend rules.)

Solution: Set `allow_downloads=True` (which is already done for a
preceding `go mod download`) to allow `go` to contact the Go module
proxy as needed by not setting `GOPROXY=off`.

Fixes #21529.