Reimplement the pants_bin_name functionality in Rust. (#21872)
This feature allows users to set a custom binary name
for use in help messages and documentation. So that
if their org has a custom wrapper script or binary,
we can direct users to use that.

Reimplementing this in Rust allows us to further 
simplify OptionsBootstrapper.

Note that the old implementation relied on setting
an env var on the current process by manipulating
`os.environ`. This served two purposes:

1) Communicating the value within the process, and
2) Communicating the value from the client to pantsd,
since the latter inherits the environment of the former.

( 2) is important because pantsd uses bin_name() at
import time, and options (including the pants_bin_name
option) may not be available yet. )

But - setting an environment variable on the current
process is not thread-safe on POSIX, and there is no
way to make it so. Therefore we change the implementation
to use a static Mutex for 1), and an explicit pass-through
of that value when spawning pantsd for 2).