perform tilde expansion on path_metadata_request
Within the pants repo itself, the inputs to `path_metadata_request` are
commonly used on the components of the `PATH` variable.  These would
presumable be valid paths, but people have all sorts of crusty
oddities in their shell configs.

In particular if one does something reasonable looking like `export
PATH="~/.local/bin:$PATH"` (note the quotes) than the path component
will be a literal `~/.local/bin` and not expanded to the home
directory.  This is because tilde expansion happened *before*
<https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_06>
the quote removal.

Since bash and friends are cool fine with literal tildes in `PATH`,
everyone would reasonable expect pants to be too.