warn instead of erroring out on almost certainly unintended PATH pref… (Cherry-pick of #21916) (#21923)
…ixes

There are a myriad reasons why one of the prefixes in PATH might not be
a "real" path: shell quoting rules are of the occult, it is easy to
copy-paste something wrong, and IDEs might inject something whacky. Most
tools (ex: bash!) will happily and silently ignore all of this, so
throwing an error is outside of norms.

The catch here of course is that a directory really could start with
`$`, `~`, or any other character. I think trying to combine relative
paths with obvious footgun names intentionally is too unlikely to merit
a configuration option.

ref #21907

Co-authored-by: cburroughs <chris.burroughs@gmail.com>