Execute or postpone most deprecation-removals scheduled for 2.26 (#21926)
We're starting the 2.26.x release series so the deprecations that were
scheduled for removal can be removed. This PR does handles of them.

I've postponed a few, since they were only relatively recently
deprecated and require very minimal code on our side, so we can be a
little more generous to any users. (e.g. the
`[GLOBAL].native_options_validation` and `[ruff]` options are unused, so
it's just some option definitions that can hang around to be deleted
later.)

There's an extra deprecation ready for removal in the Python FaaS code,
which is done in #21928 instead, because it's more fiddly/extensive.