fix CI "build wheels" step for rustup no longer installing default toolchain (Cherry-pick of #22037) (#22038)
Fix the "Build wheels" step for a breaking change in [Rustup
v1.28.0](https://blog.rust-lang.org/2025/03/02/Rustup-1.28.0.html) which
no longer installs a default toolchain any more.

Use `rustup default` to install the applicable toolchain because it is
idempotent and will install the toolchain if it is not already
installed. `rustup toolchain install` by design will try to install even
if the toolchain is already installed. See
https://github.com/rust-lang/rustup/issues/710 for discussion of this
distinction.

Closes https://github.com/pantsbuild/pants/issues/22036

---------

Co-authored-by: Tom Dyas <tom@shoalsoftware.com>