migrate to `packaging` requirements parsing from deprecated `pkg_resources` (#22033)
Migrate to use `Requirement` from `packaging` instead of the deprecated
`pkg_resources`.

The `project_name` attribute is called `name` by
`packaging.requirements.Requirement` so migrate Pants naming to use
that, except in lock files where we keep the `project_name` naming to
avoid breaking existing lockfiles.

Note in the release notes that the `.*` pattern is only permitted with
`==` and `!=` operators. `pkg_resources` accepted `.*` patterns with
other operators which is not permitted by the current versioning
specification.

---------

Co-authored-by: Benjy Weinberger <benjyw@gmail.com>