Support local path "replace" in go.mod (#22091)
This patch adds proper interpretation of a `replace` clause that lacks a
Version: interpret the dependency as a first-party module within the
same repository and not a third-party module that requires remote
fetching. This version of the `replace` clause is described at
https://go.dev/ref/mod#go-mod-file-replace
> If the path on the right side of the arrow is an absolute or relative
path
> (beginning with ./ or ../), it is interpreted as the local file path
to the
> replacement module root directory, which must contain a go.mod file.
The
> replacement version must be omitted in this case.

This patch does not add dependency inference across these references,
but merely allows the explicit dependencies to function in both pants
and go.

Fixes #14996. and probably #19463.