New `python_format_string` codegen for kubernetes (#22034)
The pr introduces a new `python_format_string` target that takes a
python format string as an input and produces the output by substituting
the variables into it. Right now it only produces `k8s_sources`, but it
can be extended in the future.

Merge after https://github.com/pantsbuild/pants/pull/22040