Switch [dockerfile-parser].use_rust_parser to true by default, deprecate non-native parser entirely (#22124)
This continues the work to switch to the native/Rust-based Dockerfile
parser by changing the default for `[dockerfile-parser].use_rust_parser`
to `true`, and deprecating setting it to `false`, i.e. deprecate using
the non-Rust parser.

Once this deprecation expires, we can remove the non-Rust parser
entirely.

This phasing work started in #21859, which was released in 2.25 just
recently. Hopefully we'll thus catch any issues with the new parser over
the next few months, before 2.27 goes out.

(This is resolving 1 out of 4 deprecations expiring in 2.27.0.dev0.
#22125 resolves the other 3.)