Delete the very-unsupported manual installation instructions (#22137)
This deletes the "manual installation" documentation page, relating to
the old `./pants` script, and updates the links to it appropriately.

Current page:
https://www.pantsbuild.org/2.25/docs/getting-started/manual-installation

Why do this? We very much do not support this script installing new
versions of Pants: we blocked this for all invocations of Pants in
#19075, which was released in 2.18 (Nov 2023). This PR will be merging
for 2.27 (likely released May or June 2024). I think the script is
unlikely to work for new versions of Pants, because it doesn't know how
to find the releases on GitHub (only PyPI), and (related to #22136) this
page now-incorrectly refers to using Python 3.7, 3.8 or 3.9.

Thus, these docs are just risking causing confusion (I don't know of a
concrete example where someone has been confused, but it's possible!).
If someone needs this info for an older version of Pants they can and
should refer to the versioned page, e.g.
https://www.pantsbuild.org/2.17/docs/getting-started/manual-installation
(note `/2.17/` in the URL).

NB. technically someone can set `NO_SCIE_WARNING=1` to silence the
blocking, but they'll need to work through a bunch of other issues (like
the GitHub Releases ones) to make that work. Reminder of relevant code:
https://github.com/pantsbuild/pants/blob/95d621a7185f1008589de027c153c614b1f99dff/src/python/pants/bin/pants_runner.py#L95-L108