Stop --keep-sandboxes from restarting the Pantsd daemon (#22127)
This turns `--keep-sandboxes=...` from a "bootstrap" option into a
normal "global" option. This means changing it doesn't require
restarting the Pantsd daemon.

This is implemented by propagating the option into every process
execution request (in the "execution environemnt"), rather than having
it associated with the long-lived local & docker command runners. Doing
this means it no longer has to be fixed when initialising the daemon,
i.e. it can become a normal global option, as desired.

The hope is that this change shortens debugging loops, e.g. `pants test
::` => "oh, weird failure" => `pants --keep-sandboxes=on_failure test
::` is much quicker without the second one having to restart the daemon.

The commits are individually sensible, building to the final feature.

Fixes #19018